# docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mind The Wait API",
        "description": "Mobile API for iOS app - Transit headway monitoring and predictions",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://localhost",
            "description": "Local development"
        },
        {
            "url": "https://mindthewait.ca",
            "description": "Production"
        }
    ],
    "paths": {
...
        "/api/v1/routes": {
            "get": {
                "tags": [
                    "Routes"
                ],
                "summary": "List routes",
                "description": "Returns all routes with performance metrics and grades",
                "operationId": "get_api_v1_routes_list",
                "parameters": [
                    {
                        "name": "city",
                        "in": "query",
                        "description": "Filter routes by city slug",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of routes with performance metrics",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "routes": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "14514"
                                                    },
                                                    "short_name": {
                                                        "type": "string",
                                                        "example": "16"
                                                    },
                                                    "long_name": {
                                                        "type": "string",
                                                        "example": "Exhibition / 8th Street"
                                                    },
                                                    "color": {
                                                        "type": "string",
                                                        "example": "FF8000"
                                                    },
                                                    "grade": {
                                                        "type": "string",
                                                        "enum": [
                                                            "A",
                                                            "B",
                                                            "C",
                                                            "D",
                                                            "F"
                                                        ],
                                                        "example": "A"
                                                    },
                                                    "on_time_pct": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 92.5
                                                    },
                                                    "active_vehicles": {
                                                        "type": "integer",
                                                        "example": 3
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "timestamp": {
                                            "type": "integer",
                                            "example": 1760939266
                                        },
                                        "city": {
                                            "type": "string",
                                            "example": "saskatoon",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/routes/{gtfsId}": {
            "get": {
                "tags": [
                    "Routes"
                ],
                "summary": "Get route details",
                "description": "Returns detailed route metadata and 30-day performance statistics",
                "operationId": "get_api_v1_route_detail",
                "parameters": [
                    {
                        "name": "gtfsId",
                        "in": "path",
                        "description": "Route GTFS ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Route details with performance stats",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "route": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "14514"
                                                },
                                                "short_name": {
                                                    "type": "string",
                                                    "example": "16"
                                                },
                                                "long_name": {
                                                    "type": "string",
                                                    "example": "Exhibition / 8th Street"
                                                },
                                                "color": {
                                                    "type": "string",
                                                    "example": "FF8000"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "stats": {
                                            "description": "30-day performance statistics",
                                            "type": "object"
                                        },
                                        "timestamp": {
                                            "type": "integer",
                                            "example": 1760940344
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Route not found"
                    }
                }
            }
        },
...
    },
...
}